Add files via upload
This webpage is part of a personal portfolio project showcasing individual details. It contains a structured layout with a navigation bar, personal information, and footer. The page presents essential personal details such as name, parents’ names, college, and full residential address. It also includes navigation links to the homepage and education details page, along with a “Go Back” button for easy navigation. Designed with simplicity and clarity, this page serves as a personal profile section within a larger portfolio website.